% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{append_stopwords}
\alias{append_stopwords}
\title{Append Stopwords}
\usage{
append_stopwords(data_stopwords, stopwords_to_append)
}
\arguments{
\item{data_stopwords}{A character vector containing the initial set of stopwords to which more stopwords can be appended.}

\item{stopwords_to_append}{A character vector containing additional stopwords to be appended to the existing \code{data_stopwords} vector.}
}
\value{
Returns a character vector of unique stopwords after appending the specified stopwords and additional stopwords from the \code{stopwords} package.
}
\description{
This function is used to append additional stopwords to an existing list.
}
\details{
The \code{append_stopwords} function takes an existing list of stopwords (\code{data_stopwords}) and appends a new set of stopwords (\code{stopwords_to_append}) to it. Additionally, stopwords from the \code{stopwords} package (using the "dutch" language option) are also appended to the list. The final list is then returned with duplicate stopwords removed by calling \code{unique()}.
}
