% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{check_all_settings}
\alias{check_all_settings}
\title{Check all setting}
\usage{
check_all_settings(settings)
}
\arguments{
\item{settings}{A current list of settings (list)}
}
\value{
Either an error or TRUE
}
\description{
Check for every setting the type and value
}
