% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{check_setting}
\alias{check_setting}
\title{Check setting}
\usage{
check_setting(settings, setting_name, setting_value)
}
\arguments{
\item{settings}{A current list of settings (list)}

\item{setting_name}{The name of the setting that needs updating (string)}

\item{setting_value}{The value of the setting}
}
\value{
TRUE if correct, FALSE if incorrect
}
\description{
Check the value of one particular setting
}
