% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_clean.R
\name{correct_text}
\alias{correct_text}
\title{Correct Text}
\usage{
correct_text(input_text, corrections)
}
\arguments{
\item{input_text}{A character vector containing the text to be corrected.}

\item{corrections}{A data frame containing two columns: \code{typoEdit} (pattern) and \code{replacement} (correction).}
}
\value{
A character vector with corrected text.
}
\description{
Applies typo corrections to an input text using a given correction table.
}
