% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_clean.R
\name{create_dataset_narratives}
\alias{create_dataset_narratives}
\title{Create Narratives Dataset}
\usage{
create_dataset_narratives(data)
}
\arguments{
\item{data}{A data.table or data.frame containing narrative data.}
}
\value{
A structured data.table with cleaned text.
}
\description{
Restructures a dataset by normalizing and transforming its textual content.
}
