% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{default_stopwords}
\alias{default_stopwords}
\title{Load Stopwords}
\usage{
default_stopwords(stopwords_file)
}
\arguments{
\item{stopwords_file}{A string representing the path to a file containing a list of stopwords, with one word per line. This file is expected to have tab-separated values with the stopwords in the first column.}
}
\value{
Returns a character vector of stopwords read from the specified file.
}
\description{
Load stopwords from a file
}
\details{
The \code{default_stopwords} function loads a list of stopwords from a specified file, with each stopword being listed on a new line. It returns a character vector of these stopwords.
}
