% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_clean.R
\name{get_corrections}
\alias{get_corrections}
\title{Load Corrections}
\usage{
get_corrections(corrections_file)
}
\arguments{
\item{corrections_file}{Path to a CSV file with typo corrections.}
}
\value{
A data frame containing \code{typoEdit} (pattern) and \code{replacement} (correction).
}
\description{
Reads a typo correction file and prepares it for use in text cleaning.
}
