% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_clean.R
\name{replace_abbr}
\alias{replace_abbr}
\title{Replace Abbreviations}
\usage{
replace_abbr(abbr)
}
\arguments{
\item{abbr}{A character string containing the abbreviation.}
}
\value{
A corrected abbreviation string.
}
\description{
Replaces periods in abbreviations to ensure correct spacing and formatting.
}
