% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumup-package.R
\docType{package}
\name{sumup-package}
\alias{sumup}
\alias{sumup-package}
\title{sumup: Utilizing Automated Text Analysis to Support Interpretation of Narrative Feedback}
\description{
Combine topic modeling and sentiment analysis to identify individual students' gaps, and highlight their strengths and weaknesses across predefined competency domains and professional activities.
}
\author{
\strong{Maintainer}: Joyce Moonen - van Loon \email{j.moonen@maastrichtuniversity.nl} (\href{https://orcid.org/0000-0002-8883-8822}{ORCID})

}
\keyword{internal}
