% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{update_educational_framework_settings}
\alias{update_educational_framework_settings}
\title{Update settings related to the educational framework}
\usage{
update_educational_framework_settings(settings)
}
\arguments{
\item{settings}{A current list of settings (list)}
}
\value{
An updated list containing all settings used in 'sumup'
}
\description{
When changing the seeds_file or use_seeds settings, update all parameters relating to the educational framework
}
