% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{update_setting}
\alias{update_setting}
\title{Update setting}
\usage{
update_setting(settings, setting_name, new_value)
}
\arguments{
\item{settings}{A current list of settings (list)}

\item{setting_name}{The name of the setting that needs updating (string)}

\item{new_value}{The new value of the setting}
}
\value{
A list containing all settings used in 'sumup'
}
\description{
Replace the value of one particular setting value
}
